backend/requirements.txt
fastapi==0.111.1
uvicorn==0.23.1
pydantic==2.7.0
numpy==1.27.5
sqlalchemy==2.0.22
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1

